// src/client/features/playback/track/client-log/utils/types.ts
export type LiveClientLogBaseParams = {
  contentId: string;
  streamUrl?: string;
  wrapper?: LivePlayerWrapper | null;
  qualityManager?: LivePlaybackQualityManager;
};

export interface LiveClientLogBaseInfo {
  is_live: boolean;
  content_id: string;
  position?: number;
  pageType: LinearPageTypes;
  bufferLength?: number;
  contentBwEstimate?: number;
  streamToken: string;
  sessionStartTs: number;
  state?: keyof typeof LivePlayerState | 'unknown';
  stateNum?: number;
  elReadyState?: number;
  playbackId?: string;
  usingApollo?: boolean;
  isAppHidden?: boolean;
  bitrate?: number;
  playback_codec_detail?: unknown;
  cdn?: string;
  isUsingWebWorker?: boolean;
  resolution?: unknown;
  screenResolution?: unknown;
  cdnProvider?: string;
  cdnCache?: string;
  cdnCloudFrontId?: string;
  cdnCloudFrontPop?: string;
  cdnServedBy?: string;
  sdkVersion?: string;
  player_type: number;
}